affordance hierarchy
Affordances can be nested within each other by defining one or more
"nesting levels":

```bnb
place Account Creation
    Fields
    > Email
    > Password
    >> Show Password
    > Actions
    >> Sign Up
    >> Back
    Footer
```

In this example, there's a single `place` named `Account Creation`,
which has 8 `affordances`:

- `Fields` and `Footer` are at the root (0) level of the place.
- `Email`, `Password` and `Actions` are at level 1.
- `Show Password`, `Sign Up` and `Back` are at level 2.

The AST exposes this level with a new `level` field on `Affordance`.
Affordances are parsed in-order, but there are no references between
affordances, so it's up to the consumer of the AST to determine the fact
that `Show Password` is nested under `Password`, based on the order of
affordances in the `affordances` field of the `Place`.

Signed-off-by: Jean Mertz <git@jeanmertz.com>