descriptions for Places, Components and Affordances
The AST has been updated to support descriptions for `Place`,
`Component` and `Affordance`. Support for descriptions in other types
can be added in the future.

The DSL parser supports the following syntax to add descriptions:

```bnb
place Home
    include Header

    # A description for an affordance.
    Log In

component Header
    # With affordances.
    Logo
```

The parser uses the following rules to parse comments(/descriptions):

- Comment has to start with `#`, excluding prefix whitespace.
- Character `#` is stripped from the comment.
- Can be spread over multiple lines, each starting with `#`.
- Whitespace after `#` is optional, if present, one space is removed.
- Other whitespace (prefixed and suffixed) is preserved.

Signed-off-by: Jean Mertz <git@jeanmertz.com>