add `ComputedSize` plugin
The computed size of a node in a tree. The size of a node can be one of
three variants:

Pending
-------
A `Pending` computed size means the size will be known eventually, but
is waiting on other data to be generated before the final size can be
determined.

For example, an entity with the `Text` component will get the computed
size defined once `bevy_text::TextPipeline::queue_text()` runs, and the
`bevy_text::TextLayoutInfo` component is added.

Inherit
-------
An `Inherit` computed size means the size is inherited from the child
nodes. It is an invalid state for leaf entities to have a computed size
of `Inherit`.

Static
------
A `Static` computed size means the size is known for this node, without
the need to iterate into the node's children to calculate it.

For any tree of nodes, the following invariants must hold:

- Each node (entity) in the tree *MUST* have the `ComputedSize`
  component attached.

- If an inheriting node has more than one child, then those children
  *MUST* have the `Transform` component attached.

- Each leaf node in the tree *CANNOT* have an inherited computed size.

This plugin also provides an additional `Padding` component, which can
be attached to any node in the tree.

The way padding is used differs per situation:

- If a node with a static size has padding, then that padding is added
  directly to the node's computed size.

- If an inheriting node has padding, then that padding is added to the
  computed size of all combined children nodes (which can have padding
  of their own).

Signed-off-by: Jean Mertz <git@jeanmertz.com>