force redraw window on specific events
Update the `WindowPlugin` plugin to run a system that detects certain
events throughout the app lifecycle, and request a redraw from the
rendering pipeline.

This is needed, because we run the app in "reactive mode", which
avoids continuously running the update cycle if no action is happening,
but it might result in the next queued redraw not happening until
another event is triggered, while changes in the rendering pipeline are
needed to be drawn right away.

With these changes, we can force one more update cycle to trigger the
rendering pipeline, to make sure the app is rendered as expected.

Signed-off-by: Jean Mertz <git@jeanmertz.com>