support manual redraw of screen
A new `ForceRedraw` resource can be set to `true`, which forces the
window to redraw itself at the end of an update cycle.

This is useful for example, if a UI button is clicked, but the mouse
isn't moved after the click, which then keeps the button in the
"active" visual state.

By asking for a redraw, the visual state is correctly returned to
"inactive".

Signed-off-by: Jean Mertz <git@jeanmertz.com>