Differentiate between DSL comments and description-comments
This extends upon the work introduced in commit
b0f19a4049cec65615c4472627c9f202812bf95e.

It introduces a second type of comments, which are ignored by the
parser. This allows adding comments to the DSL file that are only of
interest to those reading the DSL, but should be ignored by any tool
(such as the Butter GUI) that renders the Breadboard.

To account for this change, comments are modified as follows:

- All rules introduced in the linked commit still apply, except that
  description comments now have to start with `///` (three forward
  slashes).

- Any regular comment (the ones not parsed into the AST) must start with
  `//` (two forward slashes).

- Any old comments starting with `#` are now considered invalid syntax.

Signed-off-by: Jean Mertz <git@jeanmertz.com>